Add a fixture for setting up neon_env and metrics server
